Update gpio parameters types

+++ src/gpio.c
@@ -12,11 +12,11 @@ void setup_gpio() {
   softPwmCreate(GPIO_FAN, FAN_MIN_VALUE, GPIO_PWM_RANGE);
 }
 
-void update_resistor(int value) {
+void update_resistor(float value) {
   softPwmWrite(GPIO_RESISTOR, value);
 }
 
-void update_fan(int value) {
+void update_fan(float value) {
   if (value < FAN_MIN_VALUE)
     value = FAN_MIN_VALUE;
   softPwmWrite(GPIO_FAN, value);
